move integration tests, add adapter tests, pass *url.URL instead of url.URL
